requirements.txt
discord.py
ffmpeg
yt_dlp
python-dotenv
pynacl